% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_of_day.R
\name{time_announcement}
\alias{time_announcement}
\title{Time of day}
\usage{
time_announcement(time_of_day, am_v_pm)
}
\arguments{
\item{time_of_day}{the hour in which we are in}

\item{am_v_pm}{am versus pm}
}
\description{
Time of day
}
